pyproject.toml
[project]
name = "adaptive-taxonomy-mapper"
version = "1.0.0"
description = "AI-powered inference engine for mapping user-generated content to internal taxonomy"
requires-python = ">=3.9"
dependencies = [
    "groq>=0.33.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["taxonomy_mapper"]
